Moving the tests to the controllers/tests package
