+++ controllers/tests/kafkacluster_controller_istioingress_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package controllers
+package tests
 
 import (
 	"context"
